moved exceptions
