+++ src/pint/models/absolute_phase.py
@@ -5,7 +5,8 @@ from loguru import logger as log
 
 import pint.toa as toa
 from pint.models.parameter import MJDParameter, floatParameter, strParameter
-from pint.models.timing_model import MissingParameter, PhaseComponent
+from pint.models.timing_model import PhaseComponent
+from pint.exceptions import MissingParameter
 
 
 class AbsPhase(PhaseComponent):
