+++ src/pint/models/binary_ell1.py
@@ -17,7 +17,7 @@ from pint.models.stand_alone_psr_binaries import binary_orbits as bo
 from pint.models.stand_alone_psr_binaries.ELL1_model import ELL1model
 from pint.models.stand_alone_psr_binaries.ELL1H_model import ELL1Hmodel
 from pint.models.stand_alone_psr_binaries.ELL1k_model import ELL1kmodel
-from pint.models.timing_model import MissingParameter
+from pint.exceptions import MissingParameter
 from pint.utils import taylor_horner_deriv
 from pint import Tsun
