+++ src/pint/models/dmwavex.py
@@ -6,7 +6,7 @@ from loguru import logger as log
 from warnings import warn
 
 from pint.models.parameter import MJDParameter, prefixParameter
-from pint.models.timing_model import MissingParameter
+from pint.exceptions import MissingParameter
 from pint.models.dispersion_model import Dispersion
 from pint import DMconst, dmu
