+++ src/pint/models/frequency_dependent.py
@@ -6,7 +6,8 @@ import astropy.units as u
 import numpy as np
 
 from pint.models.parameter import prefixParameter
-from pint.models.timing_model import DelayComponent, MissingParameter
+from pint.models.timing_model import DelayComponent
+from pint.exceptions import MissingParameter
 
 
 class FD(DelayComponent):
