+++ src/pint/models/glitch.py
@@ -6,8 +6,9 @@ import numpy as np
 from loguru import logger as log
 
 from pint.models.parameter import prefixParameter
-from pint.models.timing_model import MissingParameter, PhaseComponent
+from pint.models.timing_model import PhaseComponent
 from pint.utils import split_prefixed_name
+from pint.exceptions import MissingParameter
 
 
 class Glitch(PhaseComponent):
