+++ src/pint/models/ifunc.py
@@ -4,7 +4,8 @@ import astropy.units as u
 import numpy as np
 
 from pint.models.parameter import floatParameter, prefixParameter
-from pint.models.timing_model import PhaseComponent, MissingParameter
+from pint.models.timing_model import PhaseComponent
+from pint.exceptions import MissingParameter
 
 
 class IFunc(PhaseComponent):
