+++ src/pint/models/piecewise.py
@@ -4,8 +4,9 @@ import astropy.units as u
 import numpy as np
 
 from pint.models.parameter import prefixParameter
-from pint.models.timing_model import MissingParameter, PhaseComponent
+from pint.models.timing_model import PhaseComponent
 from pint.utils import split_prefixed_name, taylor_horner
+from pint.exceptions import MissingParameter
 
 
 class PiecewiseSpindown(PhaseComponent):
