+++ src/pint/models/spindown.py
@@ -6,9 +6,10 @@ import astropy.units as u
 import numpy
 
 from pint.models.parameter import MJDParameter, prefixParameter
-from pint.models.timing_model import MissingParameter, PhaseComponent
+from pint.models.timing_model import PhaseComponent
 from pint.pulsar_mjd import Time
 from pint.utils import split_prefixed_name, taylor_horner, taylor_horner_deriv
+from pint.exceptions import MissingParameter
 
 
 class SpindownBase(PhaseComponent):
