+++ src/pint/models/wavex.py
@@ -6,7 +6,8 @@ from loguru import logger as log
 from warnings import warn
 
 from pint.models.parameter import MJDParameter, prefixParameter
-from pint.models.timing_model import DelayComponent, MissingParameter
+from pint.models.timing_model import DelayComponent
+from pint.exceptions import MissingParameter
 
 
 class WaveX(DelayComponent):
